Moved static variable from header file to c file
