+++ Buildings/Resources/src/EnergyPlus/FMUEnergyPlusStructure.c
@@ -7,7 +7,80 @@
 #include "FMUEnergyPlusStructure.h"
 #include "ModelicaUtilities.h"
 
+#include <stdlib.h>
+#include <string.h>
+
+
 static unsigned int Buildings_nFMU = 0;     /* Number of FMUs */
+static struct FMUBuilding** Buildings_FMUS; /* Array with pointers to all FMUs */
+
+FMUBuilding* instantiateZone(const char* idfName, const char* weaName,
+  const char* iddName, const char* epLibName, const char* zoneName, FMUZone* zone){
+  /* Allocate memory */
+  const size_t nFMU = getBuildings_nFMU();
+  if (nFMU == 0)
+    Buildings_FMUS = malloc(sizeof(struct FMUBuilding*));
+  else
+    Buildings_FMUS = realloc(Buildings_FMUS, (nFMU+1) * sizeof(struct FMUBuilding*));
+  if ( Buildings_FMUS == NULL )
+    ModelicaError("Not enough memory in FMUZoneInit.c. to allocate array for Buildings_FMU.");
+
+  Buildings_FMUS[nFMU] = malloc(sizeof(FMUBuilding));
+  if ( Buildings_FMUS[nFMU] == NULL )
+    ModelicaError("Not enough memory in FMUZoneInit.c. to allocate array for Buildings_FMU[0].");
+  Buildings_FMUS[nFMU]->zoneNames = malloc(sizeof(char*));
+  if ( Buildings_FMUS[nFMU]->zoneNames == NULL )
+    ModelicaError("Not enough memory in FMUZoneInit.c. to allocate array for Buildings_FMUS[0]->zoneNames.");
+
+  /* Assign the fmu name */
+  Buildings_FMUS[nFMU]->name = (char*) malloc((strlen(idfName)+1) * sizeof(char));
+  if ( Buildings_FMUS[nFMU]->name == NULL )
+    ModelicaError("Not enough memory in FMUZoneInit.c. to allocate fmu name.");
+  strcpy(Buildings_FMUS[nFMU]->name, idfName);
+
+  /* Assign the weather name */
+  Buildings_FMUS[nFMU]->weather = (char*) malloc((strlen(weaName)+1) * sizeof(char));
+  if ( Buildings_FMUS[nFMU]->weather == NULL )
+    ModelicaError("Not enough memory in FMUZoneInit.c. to allocate weather name.");
+  strcpy(Buildings_FMUS[nFMU]->weather, weaName);
+
+  /* Assign the idd name */
+  Buildings_FMUS[nFMU]->idd = (char*) malloc((strlen(iddName)+1) * sizeof(char));
+  if ( Buildings_FMUS[nFMU]->idd == NULL )
+    ModelicaError("Not enough memory in FMUZoneInit.c. to allocate IDD name.");
+  strcpy(Buildings_FMUS[nFMU]->idd, iddName);
+
+  /* Assign the Energyplus library name */
+  Buildings_FMUS[nFMU]->epLib = (char*) malloc((strlen(epLibName)+1) * sizeof(char));
+  if ( Buildings_FMUS[nFMU]->epLib == NULL )
+    ModelicaError("Not enough memory in FMUZoneInit.c. to allocate IDD name.");
+  strcpy(Buildings_FMUS[nFMU]->epLib, epLibName);
+
+  /* Assign the zone name */
+  Buildings_FMUS[nFMU]->zoneNames[0] = malloc((strlen(zoneName)+1) * sizeof(char));
+  if ( Buildings_FMUS[nFMU]->zoneNames[0] == NULL )
+    ModelicaError("Not enough memory in FMUZoneInit.c. to allocate zone name.");
+  strcpy(Buildings_FMUS[nFMU]->zoneNames[0], zoneName);
+
+  Buildings_FMUS[nFMU]->nZon = 1;
+
+  Buildings_FMUS[nFMU]->zones=malloc(sizeof(FMUZone *));
+  if ( Buildings_FMUS[nFMU]->zones== NULL )
+    ModelicaError("Not enough memory in FMUZoneInit.c. to allocate zones.");
+
+  /* Assign the zone */
+  Buildings_FMUS[nFMU]->zones[0] = zone;
+
+  incrementBuildings_nFMU();
+  ModelicaMessage("*** Leaving instantiateEnergyPlusFMU.\n");
+
+  /* Return the pointer to the FMU for this EnergyPlus instance */
+  return Buildings_FMUS[nFMU];
+}
+
+FMUBuilding* getBuildingsFMU(size_t iFMU){
+  return Buildings_FMUS[iFMU];
+}
 
 void incrementBuildings_nFMU(){
   Buildings_nFMU++;
@@ -17,6 +90,9 @@ void incrementBuildings_nFMU(){
 
 void decrementBuildings_nFMU(){
   Buildings_nFMU--;
+  if (Buildings_nFMU == 0){
+     free(Buildings_FMUS);
+   }
   ModelicaFormatMessage("*** Decreased Buildings_nFMU to %zu\n", Buildings_nFMU);
   return;
 }
@@ -24,4 +100,3 @@ void decrementBuildings_nFMU(){
 unsigned int getBuildings_nFMU(){
   return Buildings_nFMU;
 }
-
